+++ CHANGELOG.md
@@ -42,6 +42,8 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Added support to set a list of specific recipients allowed for send authorizations in the marker module [#1237](https://github.com/provenance-io/provenance/issues/1237).
 * Added a new name governance proposal that allows the fields of a name record to be updated. [#1266] (https://github.com/provenance-io/provenance/pull/1266)
   The `restrict` flag has been changed to `unrestrict` in the `BindName` request and `CreateRootName` proposal.
+* Added msg to add, finalize, and activate a marker in a single request [#770](https://github.com/provenance-io/provenance/issues/770).
+
 ### Improvements
 
 * Added assess msg fees spec documentation [#1172](https://github.com/provenance-io/provenance/issues/1172).